migrating core module to internal
