+++ fiftyone/brain/internal/core/image.py
